Modified variantgen to generate the graphics pokemon.h file.
Significant refactoring of sprite and palette lookup. Minor bugfixes
throughout. There may be other calls with have the personality and
variant in the wrong order.
The correct order is:
species
variant
personality.

+++ include/pokemon.h
@@ -480,7 +480,7 @@ void OakSpeechNidoranFFreeResources(void);
 void *OakSpeechNidoranFGetBuffer(u8 bufferId);
 
 bool8 IsTrainerRival(u16 trainerId);
-u8 GenerateMonVariant(u16 species, u32 variantSeed);
+u16 GenerateMonVariant(u16 species, u32 variantSeed);
 
 extern struct SpritePalette dynamicPalette;
 extern u16 dynamicPaletteData[16];
